#14 fix:회원기능 수정
