+++ src/main/java/com/meongmory/meongmory/domain/user/controller/UserController.java
@@ -2,8 +2,8 @@ package com.meongmory.meongmory.domain.user.controller;
 
 import com.meongmory.meongmory.domain.user.dto.request.SignUpUserReq;
 import com.meongmory.meongmory.domain.user.dto.response.SignUpUserRes;
-import com.meongmory.meongmory.domain.user.service.UserServiceImpl;
-import com.meongmory.meongmory.global.payload.ErrorResponse;
+import com.meongmory.meongmory.domain.user.service.UserService;
+import com.meongmory.meongmory.global.exception.BaseRes;
 import com.meongmory.meongmory.global.response.ResponseCustom;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -17,12 +17,12 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping(value = "/users")
 @RequiredArgsConstructor
 public class UserController {
-  private final UserServiceImpl userService;
+  private final UserService userService;
 
   @Operation(summary = "회원가입", description = "회원가입을 진행한다.")
   @ApiResponses(value = {
           @ApiResponse(responseCode = "200", description = "회원가입 성공", content = {@Content(mediaType = "application/json", schema = @Schema(implementation = SignUpUserRes.class))}),
-          @ApiResponse(responseCode = "400", description = "회원가입 실패", content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))}),
+          @ApiResponse(responseCode = "400", description = "회원가입 실패", content = {@Content(mediaType = "application/json", schema = @Schema(implementation = BaseRes.class))}),
   })
   @PostMapping("/signUp")
   public ResponseCustom<SignUpUserRes> signUp(@RequestBody SignUpUserReq signUpUserReq
